Switched Mekanism's Maven Repo to its proper one
It turns out that the creators of Mekanism already have a repo for it so I switched from CurseMaven to that repo instead, and I also included the whole Mek lineup. I also changed the config's comments a little and added the repo for Ad Astra in preparation for adding compatibility with it.